+++ Good+CoreDataProperties.h
@@ -2,8 +2,8 @@
 //  Good+CoreDataProperties.h
 //  HabitRPG2
 //
-//  Created by Joel Pridgen on 11/15/16.
-//  Copyright © 2016 Joel Pridgen. All rights reserved.
+//  Created by Joel Pridgen on 5/23/17.
+//  Copyright © 2017 Joel Pridgen. All rights reserved.
 //
 
 #import "Good+CoreDataClass.h"
@@ -18,7 +18,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic) int32_t cost;
 @property (nullable, nonatomic, copy) NSString *goodName;
 @property (nullable, nonatomic, copy) NSString *note;
-@property (nonatomic) int16_t useType;
+@property (nonatomic) int32_t useType;
 
 @end
