+++ Todo+CoreDataProperties.h
@@ -2,8 +2,8 @@
 //  Todo+CoreDataProperties.h
 //  HabitRPG2
 //
-//  Created by Joel Pridgen on 11/15/16.
-//  Copyright © 2016 Joel Pridgen. All rights reserved.
+//  Created by Joel Pridgen on 5/23/17.
+//  Copyright © 2017 Joel Pridgen. All rights reserved.
 //
 
 #import "Todo+CoreDataClass.h"
@@ -15,12 +15,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSFetchRequest<Todo *> *)fetchRequest;
 
-@property (nonatomic) int16_t difficulty;
+@property (nonatomic) int32_t difficulty;
 @property (nullable, nonatomic, copy) NSDate *dueDate;
 @property (nullable, nonatomic, copy) NSDate *effectiveDate;
 @property (nullable, nonatomic, copy) NSString *note;
 @property (nullable, nonatomic, copy) NSString *todoName;
-@property (nonatomic) int16_t urgency;
+@property (nonatomic) int32_t urgency;
 
 @end
